Two significant changes: 1) Bug fix: when iter(cell0) in the segmentation, the first cell index was 1 (not 0!) since iter was calling next() right off the bat. This generated problems at pole (iCell=0) since its cell2Site=-1 shot steepest path way off.
2) Added capability to check for "vertical" overlap in PT between two
basins for checking correspondence. The idea is that an "air mass"
should persist, which is tricky when considering breaking, mixing,...
Since basin PT max-min amplitudes tend to be large, this doesn't
eliminate many candidate correspondences at all. Especially given that
thetaExtr is used for determining major correspondences, this isn't too
useful and just adds an extra (not fully justified) condition.